Updated the version number to 0.9.3.1 instead of 0.9.2.1 as was accidently done before. Upgraded to BlinzCore 0.6.0.

+++ src/main/java/org/blinz/world/Scene.java
@@ -146,7 +146,7 @@ final class Scene {
         } else {
             SpriteContainer sc = containers.pop();
             sc.sprite = sprite;
-            sc.loc.setPosition(sprite.getX(), sprite.getY(), sprite.getLayer());
+            sc.loc.setPosition(sprite.getX(), sprite.getY(), (int)sprite.getLayer());
             return sc;
         }
     }
@@ -216,7 +216,7 @@ final class Scene {
 
         SpriteContainer(CameraSprite sprite) {
             this.sprite = sprite;
-            loc.setPosition(sprite.getX(), sprite.getY(), sprite.getLayer());
+            loc.setPosition(sprite.getX(), sprite.getY(), (int)sprite.getLayer());
         }
 
         /**
